# ci: memory and performance tests in CI