… .github/workflows/benchmark-tests-hpc.yml
… .github/workflows/integration-tests-hpc.yml
… training/docs/user-guide/benchmarking.rst
… training/pyproject.toml
… training/src/anemoi/training/diagnostics/profilers.py
… training/src/anemoi/training/schemas/dataloader.py
… training/tests/integration/config/benchmark/base.yaml
… training/tests/integration/config/benchmark/ensemble_crps.yaml
… training/tests/integration/config/benchmark/gnn.yaml
… training/tests/integration/config/benchmark/graphtransformer.yaml
… training/tests/integration/config/benchmark/lam.yaml
… training/tests/integration/config/benchmark/stretched.yaml
… training/tests/integration/conftest.py
… training/tests/integration/test_benchmark.py