## Description
CI tests which run the profiler to test model performance and memory usage

## What problem does this change solve?
This will alert us to memory usage and performance regressions on new features

##  Additional notes ##

Currently we have an integration test `test_benchmark_training_cycle` which runs an o48 GT model on 1 GPU.

It tests peak memory, av throughput and av time per batch. Currently we have a 5% tolerance for av time per batch and av throughput, and memory should batch exactly.

This test should be run while a branch is in development, to test the performance implications of the branch. it should be run at least once before merge. 

When a branch is merged into main, the tests could be run again with an `update_server=True` flag. At this point that branches values would become the new values of the server. *Note* the performance could degrade at this point, but hopefully we have run the tests before this point so we are aware of these performance changes.

```
Reference benchmark results:
--------------------
avThroughputIterPerS: 3.26iter/s (commit: 8baad, date: 2025-08-05)
avTimePerBatchS: 0.25s (commit: 8baad, date: 2025-08-05)
peakMemoryMB: 638.78MB (commit: 8baad, date: 2025-08-05)
--------------------

Local benchmark results:
--------------------
avThroughputIterPerS: 3.17iter/s (commit: 50d4f, date: 2025-08-05)
avTimePerBatchS: 0.26s (commit: 50d4f, date: 2025-08-05)
peakMemoryMB: 638.78MB (commit: 50d4f, date: 2025-08-05)
--------------------

Comparing local benchmark results against reference values from the server
PASS. Local value for avThroughputIterPerS is within 5% tolerance of the reference value (3.17iter/s local vs 3.26iter/s reference)
PASS. Local value for avTimePerBatchS is within 5% tolerance of the reference value (0.26s local vs 0.25s reference)
PASS. Local value for peakMemoryMB has improved compared to the reference value (638.78MB local vs 638.78MB reference)
```


## future extensions

- test on a larger source grid (how to get around downloading large datasets? can we read from /mlx if we test on Atos?)

- test the time spent on specific components (e.g. moving data between host and device). we dont have this granularity in the profiler currently, but these features are being worked on.

- log commit and date alongside performance on the server, to create a database of performance over time. This can be included in this PR


***As a contributor to the Anemoi framework, please ensure that your changes include unit tests, updates to any affected dependencies and documentation, and have been tested in a parallel setting  (i.e., with multiple GPUs). As a reviewer, you are also responsible for verifying these aspects and requesting changes if they are not adequately addressed. For guidelines about those please refer to https://anemoi.readthedocs.io/en/latest/***

By opening this pull request, I affirm that all authors agree to the [Contributor License Agreement.](https://github.com/ecmwf/codex/blob/main/Legal/contributor_license_agreement.md)


<!-- readthedocs-preview anemoi-training start -->
----
📚 Documentation preview 📚: https://anemoi-training--403.org.readthedocs.build/en/403/

<!-- readthedocs-preview anemoi-training end -->

<!-- readthedocs-preview anemoi-graphs start -->
----
📚 Documentation preview 📚: https://anemoi-graphs--403.org.readthedocs.build/en/403/

<!-- readthedocs-preview anemoi-graphs end -->

<!-- readthedocs-preview anemoi-models start -->
----
📚 Documentation preview 📚: https://anemoi-models--403.org.readthedocs.build/en/403/

<!-- readthedocs-preview anemoi-models end -->