@cathalobrien 2025-09-01
@copilot

> Really nice contribution thanks. I think having it/s & batch time is unnecessary duplication, which may confuse. So I would propose dropping the batch time.

done